2처

+++ src/main/java/com/spring/instagram/comment/Comment.java
@@ -2,22 +2,22 @@ package com.spring.instagram.comment;
 
 public class Comment {
 
-    private int id;
+    private Long id;
     private String name;
     private String commentStr;
 
-    public Comment(int id, String name,String commentStr){
+    public Comment(Long id, String name,String commentStr){
         this.id = id;
         this.name = name;
         this.commentStr = commentStr;
     }
 
-    public int getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(int id)
+    public void setId(Long id)
     {
         this.id = id;
     }
